requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
celery>=5.3
redis>=4.5
psycopg2-binary>=2.9
pytest>=7.0
pytest-django>=4.5
pytest-mock
factory-boy
python-dotenv
uvicorn
